add 3 new tables and finish 60% of chargeTemplate module

+++ galaxy-web/src/main/java/com/wuxiu/galaxy/web/biz/vo/ChargeCalculationRuleVO.java
@@ -28,7 +28,7 @@ public class ChargeCalculationRuleVO implements Serializable {
     private String calculationTypeName;
 
     @ApiModelProperty(value = "计费单位ID", required = true)
-    private Long calculationUnitsId;
+    private Integer calculationUnitsId;
 
     @ApiModelProperty(value = "计费单位名称", required = true)
     private String calculationUnitsName;
